[megatron] Support Internvl3/Internvl3.5 (#5647)

+++ swift/llm/template/template/internvl.py
@@ -78,7 +78,7 @@ class InternvlTemplate(Template):
             pixel_values = pixel_values.to(device=device)
             vit_embeds = model.extract_feature(pixel_values).to(device=device)
             selected = (input_ids == self.processor.encode('<IMG_CONTEXT>', add_special_tokens=False)[0])
-            inputs_embeds[selected] = vit_embeds.reshape(-1, vit_embeds.shape[-1])
+            inputs_embeds[selected] = vit_embeds.reshape(-1, vit_embeds.shape[-1]).to(dtype=inputs_embeds.dtype)
         elif is_deepspeed_enabled():
             dummy_pixel_values = torch.zeros((1, 3, 32, 32), device=device, dtype=inputs_embeds.dtype)
             vit_embeds = model.extract_feature(dummy_pixel_values).to(device=device)
